+++ Modules/Intent.Modules.Application.DomainInteractions/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 1.2.2
 
 - Improvement: Introduced components (and relocated basic entity patch logic) for upcoming JSON Patch module.
+- Improvement: Improved logic for determing the request property to use for key lookups.
 
 ### Version 1.2.1
 - Fixed: Cancellation token always being added to service call, even when service did not have a cancellation token parameter.